added dry run option

+++ RemoveOSXFile.sh
@@ -1,2 +1,7 @@
 #!/bin/bash
-find . -type d -name '__MACOSX' -exec rm -rf {} +
+
+if ["$1" -eq "-n"]; then
+    find . -type d -name '__MACOSX'
+else
+    find . -type d -name '__MACOSX' -exec rm -rf {} +
+fi
